### 2021-06-05 #3333 @whedon
[PRE REVIEW]: fishRman: A Shiny R Dashboard improving Global Fishing Watch data availability
**Submitting author:** @Shyentist (<a href="http://orcid.org/0000-0002-1848-9313">Pasquale Buonomo</a>)
**Repository:** <a href="https://github.com/Shyentist/fish-r-man" target ="_blank">https://github.com/Shyentist/fish-r-man</a>
**Version:** v1.0.0
