### 2021-06-05 #3333 @Kevin-Mattheus-Moerman
@Shyentist it looks like the YAML content at the top of the paper is not as expected/standard. In particular the affiliation field is missing. Can you add it? I believe you can also leave it empty.
