### 2021-06-05 #3333 @Shyentist
@Kevin-Mattheus-Moerman thank you, I added "Independent researcher" as affiliation. Would that be okay?
