+++ paper/paper.md
@@ -12,7 +12,7 @@ tags:
 authors:
   - name: Pasquale Buonomo
     orcid: 0000-0002-1848-9313
-	affiliation: 1
+    affiliation: 1
 affiliations:
  - name: Independent researcher
    index: 1
