### 2021-06-28 #3333 @KristinaRiemer
@saskiaotto @jules32 @HeatherWelch would you be able and willing to be a reviewer for this JOSS software submission?

If you're not familiar with JOSS, these reviews happen openly on GitHub and are guided by [this checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). There's some more information on this process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). There's a link to the GitHub repo for the submission in the first comment on this issue.
