### 2021-07-12 #3467 @KristinaRiemer
Hi @Shyentist, thanks again for your JOSS submission! This is to notify you that I'm going to be on vacation July 14 - August 8. Hopefully you and the submission reviewers @jules32 and @HeatherWelch will be able to make progress on this submission's review in my absence. If anything urgent comes up during that time period, feel free to contact EiC @arfon. If you have any questions right now, please let me know.
