### 2021-07-26 #3467 @Shyentist
Hello @HeatherWelch .
1. What do you mean that the button does not work? Does it remain greyed out? Does it do nothing upon clicking? I am unable to replicate the issue.
2. As of now, no, but that would be a nice addition indeed.
