### 2021-07-27 #3467 @jules32
Hello @Shyentist,

I’ve completed my first review, thank you for submitting. Below are some comments that I think would benefit your project; they are largely about documentation in various forms.

The Shiny App launched very quickly and was very responsive, and this seems like it will be helpful for researchers wanting to query and analyze Global Fishing Watch data. I think it would benefit from more documentation in the App itself; maybe a landing page that introduces Global Fishing watch with links, describes how people might query and use their data, and describes the data with clear metadata that the user will encounter. Here it would also be appropriate to link the fisheRman Handbook PDF that describes the App’s usage.

Then, on the Query and Analysis tabs, small examples of how to use the App would be very helpful. This could be pre-loaded so that the user can see the example and then clear it to begin their own query. (My first attempts resulted in either “Your query would return more than 1000000 rows” or “Showing 0 to 0 of 0 entries” as I was trying to familiarize myself with the functionality.)

The code itself (global.R, server.R, ui.R) could benefit from more documentation so that it is more clear how all the code works together and so that it could be easier for contributors to navigate and provide any suggestions.

In the paper, I think the statement of need and the state of the field could be more detailed; are there citations and descriptions for who is using the data, how they are using it, what issues they are encountering, and how fisheRman helps? Who is the desired user, what do they have to know, how will this integrate into their workflow?

The fisheRman handbook is very useful; I think it would benefit the user if it were linked from the repo’s README, from the App’s landing page, and from the paper.
