### 2021-07-28 #3467 @Shyentist
Hello @jules32 ,
thank you for your feedback and advice.

- Embedding the handbook in the app was planned a couple of days after my submission, when I saw another app doing the same with its own instructions. I was waiting for the review process to end before doing that. Definitely to-do.
- Also, I am aware of the scarcity of comments in the code. Improving that is also definitely in the to-do list.
- Handbook linked from the repo's README: Definitely to-do.
- About the in-app examples and the information on the metadata, I would prefer adding them to the handbook, rather than having them pop-up or be pre-loaded only to be cleared by the user. That is, because I am fairly sure those would be annoying for someone who already knows how the app works, while they would be ignored by anyone who hasn't bothered reading any documentation in the first place. I will try to figure out something that impacts less on the UX. Any suggestion is appreciated.
- About the paper, I will try my best to answer those questions.

How do we proceed now? Do I make the adjustment and the review goes on from there when I do, or is there anything else with the bot?
