### 2021-07-30 #3467 @HeatherWelch
Hi @Shyentist,

I have completed my review. Nice work on making GFW data more accessible and easy to manipulate. Below are a few questions and suggestions for improvement.

- It would be good to provide more background on the GFW dataset for people who are not familiar with GFW products already. E.g. what is AIS, what are the dimensions of the data (size, extent, date range), brief introduction of GFW's neural net used to identify gear types and fishing vs non-fishing behaviors (and link to Kroodsma et al 2018), what is the difference between fishing and vessel hours, published examples of the data being used to address real-world problems.
- I found the functionality described in the PDF very straightforward and useful.
- Fishing effort is tricky as a catch all term - this dataset also includes non-fishing behaviors (vessel hours). I wonder if there's a broader term that would encompass both fishing and non-fishing behaviors, perhaps just 'AIS data'.
- Would greatly benefit from an example of real-world problem solving. There were a lot of big [news stories](https://www.latimes.com/environment/story/2020-10-18/galapagos-islands-threats-illegal-fishing-pandemic#:~:text=Crisis%20in%20the%20Galapagos%3A%20Chinese,19%20threaten%20a%20natural%20wonder&text=Earlier%20this%20summer%2C%20more%20than,waters%20off%20Peru%20and%20Chile) last year about Chinese vessels fishing the border of the Ecuadorian EEZ near the Galapagos. It would be neat to use this as a prompt to guide an analysis example.
- Is 'References' under paper.md meant to be blank?

Query tab.
-Returned table should have column names that match the filter terms, e.g. "hours" in the table should match "vessel_hours" in the filter terms, "cell_II_lat" should match "Latitude", etc.
-For the Flag filter, please link to an [ISO look up table ](https://www.nationsonline.org/oneworld/country_code_list.htm) and note that these are Alpha 3 Codes. This will be helpful for being who don't know the code of their country of interest.
-For fishing effort at 10th degree there is an option to filter by MMSI; for fishing effort at 100th degree there is an option to filter by MMSI present. It is not clear to me why or how these are different. In the handbook, it would be helpful to have more information on the difference between MMSI and MMSI present filters, and why different options are provided depending on resolution.

Analysis tab.
-It would be helpful if the date range was included somewhere in the map. When writing out a png, it's helpful to maintain information on what time-series e.g. total fishing activity is calculated.
-I am still having trouble with the spatial visualization. It occurs when I convert the current data. Sometimes it doesn't happen, sometimes it does. I just had the error occur with this order of operations:
1. run the pre-filled filter on the query tab.
2. convert the current data on the analysis tab.
3. summarize by year on the descriptive tab.
4. visualize on the spatial tab (works).
5. summarize by year and cell_II_lon on the descriptive tab.
6. visualize on the spatial tab (nothing happens).
I have a feeling the data frame is being modified during the summary, and when you collapse by longitude there are no longer meaningful lat/lon coordinates to use for mapping.
