### 2021-07-30 #3467 @arfon
> How do we proceed now? Do I make the adjustment and the review goes on from there when I do, or is there anything else with the bot?

@Shyentist – please make your updates and let us know here when you are ready for the reviewers to take another look. You can update your paper preview at any time with `@whedon generate pdf`
