### 2021-07-31 #3467 @Shyentist
@arfon Great, thank you!

@HeatherWelch Hi again, thank you for your review and the useful feedback. I am assuming the first list concerns the Handbook, not the paper, apart from the last point.

- While the description of the data for each column would be useful and will be added, I feel like a thorough description on how the dataset was created is beyond the scope of the Handbook, the paper, and the app itself. I will, however, surely link to the work of Kroodsma for more details.
- Thank you, I try my best to make it user-friendly and understandable for non-programmers.
- Correct, I will amend accordingly.
- Definitely. I will work on it.
- Yes, the bot generates the pdf with paper.md up to the "References" section, then fills in the entries with the content of paper.bib.

Query tab

- I will explain the Filter name <--> Column name matches in the Handbook, in the section where each column will be explained, but I would prefer keeping them as they are, to avoid scaring users away with filters such as "cell_ll_lon" that looks much less friendly than "Longitude".
- I love this idea, so simple and yet so needed.
- Yes, definitely.

Analysis tab

Definitely.
That is actually how the app is supposed to behave. The `Visualize` button and results are completely unrelated to the `Summarize` button and table. `Visualize` creates a latitude-longitude plot, so a map, with the spatial data provided, and is not meant to be a "plotter" for what is summarized. `Summarize`, on the other hand, is an aggregator meant to create a new table that can be downloaded as a csv for whatever need the user might have. Basically, `Visualize` does not operate on the summarized dataframe, which does not always retain lat-lon information and can reach up to 7 different dimensions, but on the original spatial data.
