### 2021-07-31 9976ef6
Select needed column instead of dropping others (clip data bug)
Instead of dropping a column we do not need and that may be named anything, I simply select the only column we do need, which is "geom".
