### 2021-08-01 359b50f
Hyperlinked List of ISO Alpha-3 Country Codes
Next to the label for the 'Flag' filter there is now a hyperlink to a list of ISO Alpha-3 Country Codes, so that users won't have to know them by heart or look them up by themselves.
