### 2021-08-02 adceb89
Added the possibility to also plot EEZ, 24 Nautical Miles zones and 12 NM Zones
All files come from marineregions, the 12nm and 24nm have been casted from polygons to lines in order to retain only the borders. This helps making the plot clearer.

I also noticed the animations of the prettyCheckboxGroup don't actually work when the group is updated, so I deleted the animations, to avoid having some animated and some not-atimated boxes.
