### 2021-08-11 f14b1b0
minor adjustments
details noticed while amending the documentation, the code will still need to be commented further and organized
