### 2021-08-14 64f672a
minor adjustments
Global layers are now plotted with a thickness of 0.5 to make them more visible at any scale. Latitude and longitude ranges in the map sidebar have also been swapped so that they are in the same order as in the Query sidebar.
