### 2021-08-15 54df966
Added Handbook Tab
GitHub does not allow the app to directly read from a hosted PDF, so I am moving the Handbook to www/doc and giving it a static name "Handbook.pdf" so I don't have to change the code at every version. The odt file is still in the same place with the same name. Since the README file linked to the PDF, I "guessed" what the new link should be, so I hope I'm right.
