### 2021-09-02 #3467 @Shyentist
@KristinaRiemer hi, I have made [some improvements](https://github.com/Shyentist/fish-r-man/issues?q=is%3Aissue+milestone%3A%22Second+JOSS+Review%22+is%3Aclosed) for what was pointed out during the first review, and I am working on the actual paper as of now. I will surely let you know in case I need anything, thank you!
