### 2021-09-06 Shyentist/fish-r-man#73 @Shyentist
Update paper according to review feedback
Following the [feedback from the first review](https://github.com/openjournals/joss-reviews/issues/3467#issuecomment-887529337), the paper must be updated. The feedback is also quoted below.

_"In the paper, I think the statement of need and the state of the field could be more detailed; are there citations and descriptions for who is using the data, how they are using it, what issues they are encountering, and how fisheRman helps? Who is the desired user, what do they have to know, how will this integrate into their workflow?"_
