+++ paper/paper.bib
@@ -146,3 +146,36 @@
   year = {2021},
   note = {Accessed: 2021-05-31}
 }
+
+@Article{carey,
+    title = {Enhancing collaboration between ecologists and computer scientists: Lessons learned and recommendations forward.},
+    author = {Cayelan C. Carey, Nicole K. Ward, Kaitlin J. Farrell, Mary E. Lofton, Arianna I. Krinos, Ryan P. McClure, Kensworth C. Subratie, Renato J. Figueiredo, Jonathan P. Doubek, Paul C. Hanson, Philip Papadopoulos, Peter Arzberger},
+    journal = {Ecosphere},
+    year = {2019},
+    volume = {10},
+    number = {5},
+    pages = {12},
+    url = {https://doi.org/10.1002/ecs2.2753},
+}
+
+@Article{mariano,
+    title = {Introducing Programming Skills for Life Science Students.},
+    author = {Diego Mariano, Pedro Martins, Lucianna Helene Santos, Raquel Cardoso de Melo-Minardi},
+    journal = {Biochemistry and Molecular Biology Education},
+    year = {2019},
+    volume = {47},
+    number = {3},
+    pages = {288-295},
+    url = {https://doi.org/10.1002/bmb.21230},
+}
+
+@Article{lai,
+    title = {Evaluating the popularity of R in ecology.},
+    author = {Jiangshan Lai, Christopher J. Lortie, Robert A. Muenchen, Jian Yang, Keping Ma},
+    journal = {Ecosphere},
+    year = {2019},
+    volume = {10},
+    number = {1},
+    pages = {12},
+    url = {https://doi.org/10.1002/ecs2.2567},
+}
