### 2021-09-06 5a7e868
updated paper's bibliography
the last 3 articles I added were cited with full name of the first author in text, without "et al."

Maybe changing the "," to "and" in the bib file will fix this.
