+++ paper/paper.bib
@@ -149,7 +149,7 @@
 
 @Article{carey,
     title = {Enhancing collaboration between ecologists and computer scientists: Lessons learned and recommendations forward.},
-    author = {Cayelan C. Carey, Nicole K. Ward, Kaitlin J. Farrell, Mary E. Lofton, Arianna I. Krinos, Ryan P. McClure, Kensworth C. Subratie, Renato J. Figueiredo, Jonathan P. Doubek, Paul C. Hanson, Philip Papadopoulos, Peter Arzberger},
+    author = {Cayelan C. Carey and Nicole K. Ward and Kaitlin J. Farrell and Mary E. Lofton and Arianna I. Krinos and Ryan P. McClure and Kensworth C. Subratie and Renato J. Figueiredo and Jonathan P. Doubek and Paul C. Hanson and Philip Papadopoulos and Peter Arzberger},
     journal = {Ecosphere},
     year = {2019},
     volume = {10},
@@ -160,7 +160,7 @@
 
 @Article{mariano,
     title = {Introducing Programming Skills for Life Science Students.},
-    author = {Diego Mariano, Pedro Martins, Lucianna Helene Santos, Raquel Cardoso de Melo-Minardi},
+    author = {Diego Mariano and Pedro Martins and Lucianna Helene Santos and Raquel Cardoso de Melo-Minardi},
     journal = {Biochemistry and Molecular Biology Education},
     year = {2019},
     volume = {47},
@@ -171,7 +171,7 @@
 
 @Article{lai,
     title = {Evaluating the popularity of R in ecology.},
-    author = {Jiangshan Lai, Christopher J. Lortie, Robert A. Muenchen, Jian Yang, Keping Ma},
+    author = {Jiangshan Lai and Christopher J. Lortie and Robert A. Muenchen and Jian Yang and Keping Ma},
     journal = {Ecosphere},
     year = {2019},
     volume = {10},
