### 2021-09-06 #3467 @Shyentist
@KristinaRiemer hi, I have a question about the paper. One of the points in jules' review was

> The fisheRman handbook is very useful; I think it would benefit the user if it were linked from the repo’s README, from the App’s landing page, and from the paper.

I was wondering, if the url for the Handbook changes after the paper, wouldn't that be an issue? What are the rules for updating the paper, if that is even possible?
