### 2021-09-13 #3467 @KristinaRiemer
Are you expecting the handbook URL to change? I think probably as long as that URL is maintained in the GitHub repository README, users will be able to find it.
