### 2021-09-13 #3467 @Shyentist
I think so. If the documentation expands beyond the Handbook, I would like to arrange it in folders as the project goes on. I think the best course of action would be to link to the repository itself, with the README always bearing the updated URL for the Handbook (and whatever else might be needed).
