+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
     orcid: 0000-0002-1848-9313
     affiliation: 1
 affiliations:
- - name: Independent researcher
+ - name: [Open-Source for Marine and Ocean Sciences (OSMOS)](https://osmos.xyz/)
    index: 1
 date: 06 September 2021
 bibliography: paper.bib
@@ -47,7 +47,7 @@ use of the R statistical language increased fivefold from 2007 to 2018 in the fi
 Designed with ease of use in mind, `fishRman` is intended for a public of researchers,
 students, managers, and stakeholders in the fields of fisheries science, life sciences, 
 and economics, with little to no proficiency in programming, data analysis, or both, who
-intend to query, download, analyze, and visualize Global Fishing Watch data on fishing effort. 
+intend to query, download, analyze, and visualize Global Fishing Watch data. 
 
 Users who can program in R may also benefit from the software to avoid writing lines of code 
 for  what has already been implemented in the dashboard, in order to focus on other aspects 
@@ -57,7 +57,8 @@ Users with a deeper understanding of statistics and fisheries science, and with 
 of the datasets, only need to get acquainted with the software, while users that are new to the 
 field can easily learn what they need to know via `fishRman`’s official instructions for use, the 
 Handbook. Regardless of the prior knowledge of the user, reading the Handbook, which is available 
-in the software itself and in the GitHub repository, is key to the correct usage of the software.
+in the software itself and in the [GitHub repository](https://github.com/Shyentist/fish-r-man), is 
+key to the correct usage of the software.
 
 The user-friendly interface [@shiny; @shinyBS; @shinyjs; @shinyWidgets] allows users to 
 easily interact with the SQL query constructor, seamlessly building [@glue; @stringi; @countrycode]
@@ -65,7 +66,7 @@ and running queries [@bigrquery; @DBI]. In a few clicks, users are able to analy
 data in several different ways, such as:
 
 - visualizing the top n-th percentile of the dataframe for any percentage [@viridis; @sf; @maps; @ggplot], 
-a key passage in assessing how fishing effort overlaps fishing stocks, protected or endangered areas, or 
+a key passage in assessing how fishing effort overlaps fishing stocks, protected or restricted areas, or 
 another country's jurisdiction. 
 - calculating the fishing effort exerted by specific countries via certain geartypes, in precise areas. This
 is vital in assessing who is fishing where, when, and how they are doing it, so that fisheries management plans can
