### 2021-09-14 964c9d5
unlinked osmos from affiliation
there were issues with the preview, maybe this will fix them
