### 2021-09-15 e802098
updated test for JOSS submission
