+++ tests/shinytest/upload_and_spatial_for_100.R
@@ -12,5 +12,6 @@ Sys.sleep(6)
 app$setInputs(cumul_distr_percent = 95)
 app$setInputs(map_rez = 0.05)
 app$setInputs(mapped_column = "Mean fishing hours")
-app$setInputs(re_visualize_button = "click")
+app$setInputs(add_layer = "eez_boundaries_v11.gpkg")
+app$setInputs(re_visualize_button = "click", timeout_ = 150000)
 app$snapshot()
