### 2021-09-16 2209fd0
solved bug when all fishing hours are 0
To put it bluntly, when the column for fishing hours (or any column by which the user is filling the map) sums up to 0, the cumulative distribution function couldn't properly understand the question "how many rows of this dataframe are needed to reach up to 0?"

Solved with if condition for the "case 0", that gives the entire length of the dataframe.
