### 2021-10-05 #3467 @jules32
Hi @Shyentist , I’ve re-reviewed and think that you have met most of the checklist required for the JOSS review at this point. You've addressed my previous comments and I think that the Handbook has improved a lot; the Example on p. 12 was especially helpful for understanding how the App could be used. The only thing that I didn’t see in the Software Paper was the “State of the Field”; perhaps https://globalfishingwatch.org/data/working-with-our-downloadable-public-data-in-r/ would be a potential citation for this section.

In the long term (beyond the scope of this JOSS review), I do think that more documentation focused on your primary audience would increase the number of users who use the App, and improve their experience with it. For example, I think the App itself would benefit greatly from some sort of text/banner on the home page that says something like “This is the fishRman app to help you explore and analyze Global Fishing Watch Data. Learn more about Global Fishing Watch and the data available for querying with these links”. I say this because I expect that the App will be the “first contact” for many fisheries researchers, who learn about it from their colleagues who have shared the https://shyentist.shinyapps.io/fish-r-man/ link. Anyone who receives this link will have not read the manual, so having a title and information about what is up top will be very helpful. And one last thought – I have recently learned about https://documentation.divio.com/ and found it to be really helpful for thinking about writing software documentation and am sharing it here in case it’s useful for you too.

Cheers,
Julie
