### 2021-10-05 #3467 @HeatherWelch
Hi @Shyentist, I have also completed my 2nd review. Nice job on revisions, you have addressed all of my comments. I particularly like the in-depth Galapagos example in the handbook (really nicely done), and that the handbook is now directly embedded in the app. I have no further edits. Best, Heather
