### 2021-10-06 #3467 @Shyentist
Thank you @jules32 and @HeatherWelch for the 2nd review. I'll add an explanatory banner redirecting to the proper documentation and amend the state of the field as per suggestion. The link provided is, in fact, the one I often refer to when met with an issue with GFW data. About the documentation, thank you for Divio, I'll look into that, although I started reading [{officedown}](https://www.r-bloggers.com/2021/09/gxp-compliance-in-pharma-made-easier-good-documentation-practices-with-r-markdown-and-officedown/), to solve the same issue. I'll see which fares better.

Cheers 😄
