### 2021-10-07 Shyentist/fish-r-man#79 @Shyentist
Add a banner to the first page
Following the [second review for JOSS](https://github.com/openjournals/joss-reviews/issues/3467), one useful feedback is to add a banner to the first page, since fishRman will likely  be the “first contact” for many fisheries researchers, who learn about it from their colleagues who have shared the shinyapps link (not the repo with the documentation).

Something like "This is the fishRman app to help you explore and analyze Global Fishing Watch Data. Learn more about Global Fishing Watch and the data available for querying with these links"

Preferably, clicking on "these links" should open a pop-up as the ones used for Warnings, with a list of links.
