### 2021-10-07 Shyentist/fish-r-man#80 @Shyentist
Detail state of the field in JOSS paper
Following the [second review for JOSS](https://github.com/openjournals/joss-reviews/issues/3467#issuecomment-934003083), one useful feedback is to detail what the current situation is for people who would like to use GFW data. Currently, the only official guide for use seems to be [this GFW webpage](https://globalfishingwatch.org/data/working-with-our-downloadable-public-data-in-r/), specifically for an R proficient target.
