+++ paper/paper.bib
@@ -144,7 +144,7 @@
   author = {{Global Fishing Watch}},
   url = {https://globalfishingwatch.org/},
   year = {2021},
-  note = {Accessed: 2021-05-31}
+  note = {Accessed: 2021-10-08}
 }
 
 @Article{carey,
@@ -178,4 +178,28 @@
     number = {1},
     pages = {12},
     url = {https://doi.org/10.1002/ecs2.2567},
+}
+
+@misc{ClavelleR,
+  title = {Working with our downloadable public data in R},
+  author = {{Tyler Clavelle}},
+  url = {https://globalfishingwatch.org/data/working-with-our-downloadable-public-data-in-r/},
+  year = {2019},
+  note = {Accessed: 2021-10-08}
+}
+
+@misc{ClavelleGEE,
+  title = {Working with our public data: Google Earth Engine},
+  author = {{Tyler Clavelle}},
+  url = {https://globalfishingwatch.org/data/working-with-our-public-data-google-earth-engine/},
+  year = {2019},
+  note = {Accessed: 2021-10-08}
+}
+
+@misc{Mayorga,
+  title = {Getting Global Fishing Watch Data from Google Big Query using R},
+  author = {{Juan S. Mayorga}},
+  url = {http://jsmayorga.com/post/getting-global-fishing-watch-from-google-bigquery-using-r/},
+  year = {2018},
+  note = {Accessed: 2021-10-08}
 }
