### 2021-10-07 e3aeb6c
amended one test
sometimes 30 seconds are not enough to do the calculations, so I brought it to 60
