### 2021-10-08 5bf4039
edited paper and bibliography slightly
Added the DOI fields for 3 references. I thought the url type that included the DOI was automatically understood as such, but there is a specific field for it.
