+++ paper/paper.bib
@@ -56,6 +56,7 @@
     number = {43},
     pages = {1686},
     doi = {10.21105/joss.01686},
+	url = {https://doi.org/10.21105/joss.01686},
 }
   
 @Manual{glue,
@@ -136,6 +137,7 @@
     volume = {3},
     number = {28},
     pages = {848},
+	doi = {10.21105/joss.00848},
     url = {https://doi.org/10.21105/joss.00848},
 }
 
@@ -155,6 +157,7 @@
     volume = {10},
     number = {5},
     pages = {12},
+	doi = {10.1002/ecs2.2753}
     url = {https://doi.org/10.1002/ecs2.2753},
 }
 
@@ -166,6 +169,7 @@
     volume = {47},
     number = {3},
     pages = {288-295},
+	doi = {10.1002/bmb.21230}
     url = {https://doi.org/10.1002/bmb.21230},
 }
 
@@ -177,6 +181,7 @@
     volume = {10},
     number = {1},
     pages = {12},
+	doi = {10.1002/ecs2.2567}
     url = {https://doi.org/10.1002/ecs2.2567},
 }
 
