+++ paper/paper.bib
@@ -157,7 +157,7 @@
     volume = {10},
     number = {5},
     pages = {12},
-	doi = {10.1002/ecs2.2753}
+	doi = {10.1002/ecs2.2753},
     url = {https://doi.org/10.1002/ecs2.2753},
 }
 
@@ -169,7 +169,7 @@
     volume = {47},
     number = {3},
     pages = {288-295},
-	doi = {10.1002/bmb.21230}
+	doi = {10.1002/bmb.21230},
     url = {https://doi.org/10.1002/bmb.21230},
 }
 
@@ -181,7 +181,7 @@
     volume = {10},
     number = {1},
     pages = {12},
-	doi = {10.1002/ecs2.2567}
+	doi = {10.1002/ecs2.2567},
     url = {https://doi.org/10.1002/ecs2.2567},
 }
 
