### 2021-10-08 #3467 @Shyentist
Apparently, the curly brackets within curly brackets keeps the elements as they are, so I also have to edit the author fields for Tyler Clavelle and Mayorga. They are cited in text by full name and, in the references, are in alphabetical order by name rather than surname.
