### 2021-10-08 0b311fd
other minor edits to paper and bibliography
