+++ paper/paper.bib
@@ -114,7 +114,7 @@
 }
 
 @Manual{shinyBS,
-    title = {shinyBS: Twitter Bootstrap Components for Shiny},
+    title = {{shinyBS: Twitter Bootstrap Components for Shiny}},
     author = {Eric Bailey},
     year = {2015},
     note = {R package version 0.61},
@@ -122,7 +122,7 @@
 }
 
 @Manual{dplyr,
-    title = {dplyr: A Grammar of Data Manipulation},
+    title = {{dplyr: A Grammar of Data Manipulation}},
     author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
     year = {2021},
     note = {R package version 1.0.6},
@@ -130,7 +130,7 @@
 }
 
 @Article{countrycode,
-    title = {countrycode: An R package to convert country names and country codes},
+    title = {{countrycode: An R package to convert country names and country codes}},
     author = {Vincent Arel-Bundock and Nils Enevoldsen and CJ Yetman},
     journal = {Journal of Open Source Software},
     year = {2018},
@@ -142,7 +142,7 @@
 }
 
 @misc{GFW,
-  title = {Global Fishing Watch data},
+  title = {{Global Fishing Watch data}},
   author = {{Global Fishing Watch}},
   url = {https://globalfishingwatch.org/},
   year = {2021},
@@ -150,7 +150,7 @@
 }
 
 @Article{carey,
-    title = {Enhancing collaboration between ecologists and computer scientists: Lessons learned and recommendations forward.},
+    title = {{Enhancing collaboration between ecologists and computer scientists: Lessons learned and recommendations forward.}},
     author = {Cayelan C. Carey and Nicole K. Ward and Kaitlin J. Farrell and Mary E. Lofton and Arianna I. Krinos and Ryan P. McClure and Kensworth C. Subratie and Renato J. Figueiredo and Jonathan P. Doubek and Paul C. Hanson and Philip Papadopoulos and Peter Arzberger},
     journal = {Ecosphere},
     year = {2019},
@@ -162,7 +162,7 @@
 }
 
 @Article{mariano,
-    title = {Introducing Programming Skills for Life Science Students.},
+    title = {{Introducing Programming Skills for Life Science Students.}},
     author = {Diego Mariano and Pedro Martins and Lucianna Helene Santos and Raquel Cardoso de Melo-Minardi},
     journal = {Biochemistry and Molecular Biology Education},
     year = {2019},
@@ -174,7 +174,7 @@
 }
 
 @Article{lai,
-    title = {Evaluating the popularity of R in ecology.},
+    title = {{Evaluating the popularity of R in ecology.}},
     author = {Jiangshan Lai and Christopher J. Lortie and Robert A. Muenchen and Jian Yang and Keping Ma},
     journal = {Ecosphere},
     year = {2019},
@@ -186,24 +186,24 @@
 }
 
 @misc{ClavelleR,
-  title = {Working with our downloadable public data in R},
-  author = {{Tyler Clavelle}},
+  title = {{Working with our downloadable public data in R}},
+  author = {Tyler Clavelle},
   url = {https://globalfishingwatch.org/data/working-with-our-downloadable-public-data-in-r/},
   year = {2019},
   note = {Accessed: 2021-10-08}
 }
 
 @misc{ClavelleGEE,
-  title = {Working with our public data: Google Earth Engine},
-  author = {{Tyler Clavelle}},
+  title = {{Working with our public data: Google Earth Engine}},
+  author = {Tyler Clavelle},
   url = {https://globalfishingwatch.org/data/working-with-our-public-data-google-earth-engine/},
   year = {2019},
[…]
