### 2021-10-10 c83f3f6
attempted to fix bibliography
Some JOSS paapers have 3 curly brackets for their titles, so maybe that fixes the capitalization problem for me
