+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @Manual{R,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {{{R: A Language and Environment for Statistical Computing}}},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
@@ -8,7 +8,7 @@
 }
 
 @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {{{shiny: Web Application Framework for R}}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
     note = {R package version 1.6.0},
@@ -16,7 +16,7 @@
 }
 
 @Manual{shinyjs,
-    title = {shinyjs: Easily Improve the User Experience of Your Shiny Apps in Seconds},
+    title = {{{shinyjs: Easily Improve the User Experience of Your Shiny Apps in Seconds}}},
     author = {Dean Attali},
     year = {2020},
     note = {R package version 2.0.0},
@@ -24,7 +24,7 @@
 }
 
 @Manual{shinyWidgets,
-    title = {shinyWidgets: Custom Inputs Widgets for Shiny},
+    title = {{{shinyWidgets: Custom Inputs Widgets for Shiny}}},
     author = {Victor Perrier and Fanny Meyer and David Granjon},
     year = {2021},
     note = {R package version 0.6.0},
@@ -32,7 +32,7 @@
 }
 
 @Manual{DBI,
-    title = {DBI: R Database Interface},
+    title = {{{DBI: R Database Interface}}},
     author = {{R Special Interest Group on Databases (R-SIG-DB)} and Hadley Wickham and Kirill Müller},
     year = {2021},
     note = {R package version 1.1.1},
@@ -40,7 +40,7 @@
 }
   
 @Manual{bigrquery,
-    title = {bigrquery: An Interface to Google's 'BigQuery' 'API'},
+    title = {{{bigrquery: An Interface to Google's 'BigQuery' 'API'}}},
     author = {Hadley Wickham and Jennifer Bryan},
     year = {2020},
     note = {R package version 1.3.2},
@@ -48,7 +48,7 @@
 }  
 
 @Article{tidyverse,
-    title = {Welcome to the {tidyverse}},
+    title = {{{Welcome to the tidyverse}}},
     author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
     year = {2019},
     journal = {Journal of Open Source Software},
@@ -60,7 +60,7 @@
 }
   
 @Manual{glue,
-    title = {glue: Interpreted String Literals},
+    title = {{{glue: Interpreted String Literals}}},
     author = {Jim Hester},
     year = {2020},
     note = {R package version 1.4.2},
@@ -68,7 +68,7 @@
 }  
   
 @Manual{stringi,
-    title = {R package stringi: Character string processing facilities},
+    title = {{{R package stringi: Character string processing facilities}}},
     author = {Marek Gagolewski},
     year = {2020},
[…]
