### 2021-10-12 1f231e1
attempted capitalization
only some capital letters have been case protected (double curly brackets) to check whether they differ from the others
