+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @Manual{R,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {{R}: {A} {L}anguage and {E}nvironment for {S}tatistical {C}omputing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
@@ -130,7 +130,7 @@
 }
 
 @Article{countrycode,
-    title = {countrycode: {{An R}} package to convert country names and country codes},
+    title = {countrycode: {A}n {R} package to convert country names and country codes},
     author = {Vincent Arel-Bundock and Nils Enevoldsen and CJ Yetman},
     journal = {Journal of Open Source Software},
     year = {2018},
