### 2021-10-12 a61c73b
deleted one curly brackets from bibliography
