+++ paper/paper.bib
@@ -104,7 +104,7 @@
 }
 
 @Manual{viridis,
-    title = {viridis} - Colorblind-Friendly Color Maps for R},
+    title = {viridis - Colorblind-Friendly Color Maps for R},
     author = {{Garnier} and {Simon} and {Ross} and {Noam} and {Rudis} and {Robert} and {Camargo} and Antônio Pedro and {Sciaini} and {Marco} and {Scherer} and {Cédric}},
     year = {2021},
     note = {R package version 0.6.1},
