### 2021-10-12 82e6509
probably solved capitalization issue in bibliography
