### 2021-10-12 09f66ca
amended bibliography
