+++ paper/paper.bib
@@ -48,7 +48,7 @@
 }  
 
 @Article{tidyverse,
-    title = {{Welcome to the tidyverse}},
+    title = {{Welcome to the Tidyverse}},
     author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
     year = {2019},
     journal = {Journal of Open Source Software},
@@ -71,7 +71,7 @@
     title = {{R package stringi: Character string processing facilities}},
     author = {Marek Gagolewski},
     year = {2020},
-    url = {http://www.gagolewski.com/software/stringi/},
+    url = {https://stringi.gagolewski.com/},
 }  
   
 @Article{sf,
@@ -88,14 +88,14 @@
   
 @Manual{maps,
     title = {{maps: Draw Geographical Maps}},
-    author = {Original S code by Richard A. Becker and Allan R. Wilks. R version by Ray Brownrigg. Enhancements by Thomas P Minka and Alex Deckmyn.},
+    author = {Richard A. Becker and Allan R. Wilks and Ray Brownrigg and Thomas P. Minka and Alex Deckmyn},
     year = {2018},
     note = {R package version 3.3.0},
     url = {https://CRAN.R-project.org/package=maps},
 }  
   
 @Book{ggplot,
-    author = {Hadley Wickham}},
+    author = {Hadley Wickham},
     title = {{ggplot2: Elegant Graphics for Data Analysis},
     publisher = {Springer-Verlag New York},
     year = {2016},
@@ -105,7 +105,7 @@
 
 @Manual{viridis,
     title = {{viridis - Colorblind-Friendly Color Maps for R}},
-    author = {{Garnier} and {Simon} and {Ross} and {Noam} and {Rudis} and {Robert} and {Camargo} and Antônio Pedro and {Sciaini} and {Marco} and {Scherer} and {Cédric}},
+    author = {Simon Garnier and Noam Ross and Robert Rudis and Antônio P. Camargo and Marco Sciaini and Cédric Scherer},
     year = {2021},
     note = {R package version 0.6.1},
     url = {https://sjmgarnier.github.io/viridis/},
