### 2021-10-12 e17a5c5
added closing curly bracket to bib file
