+++ paper/paper.bib
@@ -96,7 +96,7 @@
   
 @Book{ggplot,
     author = {Hadley Wickham},
-    title = {{ggplot2: Elegant Graphics for Data Analysis},
+    title = {{ggplot2: Elegant Graphics for Data Analysis}},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
