### 2021-10-14 #3467 @KristinaRiemer
Thanks for making all of those changes @Shyentist! I'm actually surprised that so many of the citations from `citations` had errors and such, I would think we should be able to trust that they'll be correct.

I think the paper looks good now. The next steps are to: 1) [create a tagged release](https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release) of the repo on GitHub and 2) archive the code (in Zenodo, figshare, institutional repository, etc.) to generate a DOI. If you can post a link to the DOI here when you're done please.
