### 2021-10-19 #3467 @KristinaRiemer
@Shyentist the DOI in the generated PDF will be the one for the published paper. It's actually generated from the review issue #, which is 3467 for your submission. I wasn't actually sure about how DOIs were created, I had to ask someone else, so thanks for bringing this up! Feel free to add that DOI to the app and then do the release and archive.
