### 2021-10-20 b826c00
improved references at start and in the footer
