+++ ui.R
@@ -408,16 +408,17 @@ ui <- fluidPage(
         column(
           6,
           tags$p("References"),
-          "Software by 'Buonomo Pasquale. [2021].",
-          tags$a("https://github.com/Shyentist/fish-r-man'",
+          "Software: Buonomo P. [2021]. fishRman: A Shiny R Dashboard improving Global Fishing Watch data availability. Journal of Open Source Software.",
+          tags$a("https://doi.org/10.21105/joss.03467",
             target = "_blank",
             rel = "noreferrer noopener",
-            href = "https://github.com/Shyentist/fish-r-man",
+            href = "https://doi.org/10.21105/joss.03467",
             style = "color:#000000"
           ),
           tags$br(),
-          "Data by 'Global Fishing Watch. [2021].",
-          tags$a("https://globalfishingwatch.org/'",
+          tags$br(),
+          "Data: Global Fishing Watch. [2021].",
+          tags$a("https://globalfishingwatch.org/",
             target = "_blank",
             rel = "noreferrer noopener",
             href = "https://globalfishingwatch.org/",
