### 2021-10-20 795540e
added paper ref to README
