### 2021-10-27 bd5d2b5
created sql.construct function
This way, separation of concerns is clearer. Also, the 'glue' library is no longer needed.
