### 2021-11-13 01403b7
modularised some functions and added google authentication
I have started separating function so that I can later move to Unit Testing instead of tests by screenshot. This way, I can automate them.

Also, depending on which of two lines in global is commented in/out, the app will require the user to authenticate. This is necessary to have a downloadable version of fishRman where users rely on their Google data and RAM.
