### 2021-11-20 671524c
made stringi package not necessary
Now sub() is used, with an easy regex to point at the last "OR" of the query. This way, we do not need the entire stringi package just for that.
