### 2021-11-20 0cc5be5
separated specific case of summarize function
to keep the code tidy and because it was repeated 2 times
