### 2021-11-23 c5f0fee
made a function for the footer
