### 2021-11-23 e076bac
edited ui to accommodate new footer() function
