+++ server.R
@@ -8,7 +8,7 @@ server <- function(input, output, session) {
     
     table_name_ui <- input$table_name_ui
 
-    # check which table the user decided to query, and presents the checkboxes
+    # checks which table the user decided to query, and presents the checkboxes
     # of the fields (columns) of that table
     
     if (table_name_ui == "fishing_effort_byvessel_v2"){
@@ -51,34 +51,93 @@ server <- function(input, output, session) {
     ignoreNULL = FALSE
   )
 
+  # this function takes different paths depending on the input
+  # if "Filter" button is pressed, it creates an SQL Query runs it
+  # if csv or gpkg files are uploaded, it assign the dataframe to "my_data"
+  
+  SQL <- reactive({
+    
+    table <- input$table_name_ui
+    
+    table_full_name <- paste(
+      project,
+      dataset,
+      table,
+      sep = "."
+    )
+    
+    # Below, I check which filter boxes are checked, and assign the input
+    # value to a variable named exactly like the input's id. If the checkbox
+    # is unchecked, I assign NULL to the respective variable.
+    # I do this because, otherwise, one could i.e. select the 100th degree
+    # table, which has a flag and a geartype column, check those boxes,
+    # input something and then switch to the 10th degree table, which does
+    # not have the columns. The inputs would only be grayed out, not deleted,
+    # so the query would have them all the same.
+    
+    checked_boxes <- input$filter_columns_ui
+    
+    for (id in list_togglable_ids) {
+      if (id %in% checked_boxes) {
+        assign(id, input[[id]])
+      } else {
+        assign(id, NULL)
+      }
+    }
+    
+    # start the SQL constructor
+    SQL <- sql.construct(
+      table = table_full_name,
+      date = date,
+      cell_ll_lat = cell_ll_lat,
+      cell_ll_lon = cell_ll_lon,
+      hours = hours,
+      fishing_hours = fishing_hours,
+      mmsi_present = mmsi_present,
+      flag = flag,
+      geartype = geartype,
+      mmsi = mmsi
+    )  
+    
+    return(SQL)
+    
+  })
+  
+  observe({
+    
+    SQL <- SQL()
+    
+    output$sql_query <- renderText({
+      SQL
+    })
+    
+  })
+  
   # what is happening down here, with which_event, you will see it often in this
   # code. It is basically a switch, that I can later check in reactive expressions
   # that must differentiate among the various types of inputs that they can receive
   # for instance, both when uploading a csv or querying data, I need the final
   # result to be assigned to 'my_data', so that later on I can just refer to this
   # value, which is qualitatively the same for both sources
-
+  
   which_event <- reactiveValues(
     query = FALSE, # activated when new data is queried
-    csv = FALSE, # activated when new data is uploaded via csv
-    gpkg = FALSE # activated when new data is uploaded via gpkg
+    csv = FALSE # activated when new data is uploaded via csv
   )
-
+  
   observeEvent(input$filter_button, {
     which_event$query <- TRUE
     which_event$csv <- FALSE
-    which_event$gpkg <- FALSE
   })
-
+  
   observeEvent(input$uploaded_csv, {
     which_event$query <- FALSE
     which_event$csv <- TRUE
-    which_event$gpkg <- FALSE
   })
-
-  # the switch for which_event$gpkg is in sf_data(), when the spatial dataframe
+  
+  # the switch for which_event$gpkg is in my_data(), when the spatial dataframe
   # passes the checks and is successfully loaded
-
+  
   # I used the possibleInputs list to have reactive events start at any
   # change of the list, then check which of the previously mentioned
   # switches was turned on (TRUE)
@@ -86,89 +145,47 @@ server <- function(input, output, session) {
[…]
