### 2021-11-25 0427df5
added more code commentary
