+++ server.R
@@ -2,8 +2,6 @@ server <- function(input, output, session) {
 
   intro.message()
 
-  #depending on the chosen table, update the input checkboxes
-  
   observeEvent(input$table_name_ui, {
     
     table_name_ui <- input$table_name_ui
@@ -51,9 +49,8 @@ server <- function(input, output, session) {
     ignoreNULL = FALSE
   )
 
-  # this function takes different paths depending on the input
-  # if "Filter" button is pressed, it creates an SQL Query runs it
-  # if csv or gpkg files are uploaded, it assign the dataframe to "my_data"
+  # based on chosen table and input values to query
+  # returns an SQL query string
   
   SQL <- reactive({
     
@@ -103,6 +100,8 @@ server <- function(input, output, session) {
     
   })
   
+  # the SQL query string is rendered as text in the UI
+  
   observe({
     
     SQL <- SQL()
@@ -113,12 +112,12 @@ server <- function(input, output, session) {
     
   })
   
-  # what is happening down here, with which_event, you will see it often in this
-  # code. It is basically a switch, that I can later check in reactive expressions
-  # that must differentiate among the various types of inputs that they can receive
-  # for instance, both when uploading a csv or querying data, I need the final
-  # result to be assigned to 'my_data', so that later on I can just refer to this
-  # value, which is qualitatively the same for both sources
+  # what is happening down here, with which_event, is basically a switch, that I 
+  # can later check in reactive expressions that must differentiate among the 
+  # various types of inputs that they can receive. For instance, both when 
+  # uploading a csv or querying data, I need the final result to be assigned to 
+  # 'my_data', so that later on I can just refer to this value, which is 
+  # qualitatively the same for both sources
   
   which_event <- reactiveValues(
     query = FALSE, # activated when new data is queried
@@ -135,9 +134,6 @@ server <- function(input, output, session) {
     which_event$csv <- TRUE
   })
   
-  # the switch for which_event$gpkg is in my_data(), when the spatial dataframe
-  # passes the checks and is successfully loaded
-  
   # I used the possibleInputs list to have reactive events start at any
   # change of the list, then check which of the previously mentioned
   # switches was turned on (TRUE)
@@ -149,6 +145,10 @@ server <- function(input, output, session) {
     )
   })
 
+  # this function takes different paths depending on the input. If "Filter" button 
+  # is pressed, it runs the SQL Query created by the user and assigns the data
+  # to "my_data", if a csv file is uploaded, it assigns the dataframe to "my_data"
+  
   my_data <- eventReactive(possibleInputs(), {
     tryCatch(
       {
@@ -184,8 +184,12 @@ server <- function(input, output, session) {
           billing <- input$billing
           
           if (billing == ""){
-            billing <- "fish-r-man"
+            billing <- "fish-r-man" 
           }
+          
+          # billing defaults to fish-r-man to make the online experience more comfortable
+          # as the user is not meant to access their google account nor input 
+          # their billing project
           
           BQ_connection <- dbConnect(bigquery(),
                                      project = project,
@@ -249,6 +253,7 @@ server <- function(input, output, session) {
           # a csv with the same col_names and different types of data
           # to crash the app for the sake of it, but no one else should
           # have any issue
+          
           if ((isFALSE(all.equal(col_names_df, column_100th))) && (isFALSE(all.equal(col_names_df, column_10th)))) {
             df <- NULL
           }
@@ -264,10 +269,12 @@ server <- function(input, output, session) {
         sdf <- st_as_sf(
           df, 
           coords = c("cell_ll_lon", "cell_ll_lat"), 
-          remove = FALSE
[…]
