### 2021-12-03 cf14cfb
improved sql constructor for list inputs (IN instead of list of OR)
Instead of, for instance, "AND (flag == 'USA' OR flag == 'ITA' OR flag == 'FRA')"

Now we have a much more readable "AND flag IN ('USA', 'ITA', 'FRA')"
