+++ fun/sql.construct.R
@@ -2,8 +2,8 @@
 # SELECT * FROM {'table'}. For inputs of type numeric range
 # the query adds AND, as in "WHERE cell_ll_lat >= 10 AND cell_ll_lat < 20 ", since the
 # two things must be true at the same time. For categories, such as flag and
-# geartype, I add OR, as in " flag = 'ITA' OR flag = 'FRA' ", since we are
-# looking for entries that match any of these categories
+# geartype, I use "AND flag IN ()" with a string version of the flag input list
+# The same as flag goes for geartype
 
 # for MMSI, the SQL created is a simple 'like', but one can use the '%'
 # character to search things starting/ending/containing the input
@@ -85,27 +85,14 @@ sql.construct <- function(table = NULL, date = NULL, cell_ll_lat = NULL, cell_ll
     list_values <- unlist(list_values, recursive = TRUE, use.names = TRUE)
     
     if (all(!is.null(list_values)) && all(!is.na(list_values))) {
-      next_list_SQL <- "AND ("
       
-      for (element in list_values) {
-        next_element_SQL <- sprintf(
-          "%s = '%s' OR ",
-          list_name,
-          element
-        )
-        
-        next_list_SQL <- paste(
-          next_list_SQL,
-          next_element_SQL,
-          sep = ""
-        )
-      }
+      string_values <- paste(list_values, collapse = "', '")
       
-      next_list_SQL <- sub(
-        " OR $",
-        ") ",
-        next_list_SQL
-      )
+      next_list_SQL <- sprintf(
+        "AND %s IN ('%s')",
+        list_name,
+        string_values
+        )
       
       SQL <- paste(
         SQL,
