### 2021-12-25 b1ec302
summaries table now disappears when new data is loaded
also, the download analyses button is disabled
