### 2021-12-25 1232d19
fixed aggregation by new resolution for data (100th degree)
I was still using lat & lon as variable names, but they switched to cell_ll_lat and cell_ll_lon some commits ago
