### 2021-12-25 464fc9a
map axis are now labeled Latitude and Longitude
to maintain consistency with Front-end nomenclature used
